requirements.txt
asttokens==2.4.1
bcrypt==4.1.3
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.7
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.0.0
six==1.16.0
SQLAlchemy==2.0.30
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
